Move Operation::{generate_redos, can_merge, and merge} to to backend.rs
